Update API spec with motion spec, and add support for floating point scroll speeds

+++ src/graphics_component/motion.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct MotionConfig {
     pub direction_degrees: u16,
-    pub distance_per_tick: u32,
+    pub distance_per_tick: f32,
     pub periodicity: u32,
 }
